+++ src/components/Carousel/index.ts
@@ -1,2 +1,6 @@
 export { default } from "./Carousel";
-export type { Props as CarouselProps, InstanceRef as CarouselInstanceRef } from "./Carousel.types";
+export type {
+	Props as CarouselProps,
+	Instance as CarouselInstanceRef,
+	Instance as CarouselInstance,
+} from "./Carousel.types";