+++ src/components/_private/Flyout/index.ts
@@ -2,7 +2,7 @@ export { default } from "./Flyout";
 export { useFlyoutContext } from "./Flyout.context";
 export type {
 	Props as FlyoutProps,
-	InstanceRef as FlyoutInstanceRef,
+	Instance as FlyoutInstance,
 	TriggerProps as FlyoutTriggerProps,
 	ContentProps as FlyoutContentProps,
 } from "./Flyout.types";